autoconfigure working idempotently

+++ lib/provisioners/vagrant_box_ruby_installer.rb
@@ -1,18 +1,14 @@
 module Servitor
 
-  class VagrantBoxRubyProvisioner < VagrantBoxProvisioner
+  class VagrantBoxRubyInstaller
     attr_reader :requirements
 
     def initialize(requirements)
-      super(VagrantBaseboxProvisioner.new(requirements))
       @requirements = requirements
     end
 
-    def name
-      "#{base_provisioner.name}_ruby-#{@requirements.ruby_version}".gsub(/[^a-zA-Z0-9\-]/, '-')
-    end
-
-    def setup(box)
+    def install(box)
+      return if already_installed?(box)
       box.ssh('which curl || sudo apt-get install curl -y')
       case @requirements.ruby_version
       when /jruby/
@@ -22,12 +18,17 @@ module Servitor
       else
         box.ssh('sudo apt-get install build-essential openssl libreadline6 libreadline6-dev curl git-core zlib1g zlib1g-dev libssl-dev libyaml-dev libsqlite3-dev sqlite3 libxml2-dev libxslt-dev autoconf libc6-dev ncurses-dev automake libtool bison subversion pkg-config -y')
       end
-      box.ssh('which rvm || curl -L https://get.rvm.io | bash -s stable; echo "avoiding RVM exit code 1"')
+      box.ssh('which rvm || curl -L https://get.rvm.io | bash -s stable', :ignore_exit_code => true)
       box.ssh('which curl || sudo apt-get install curl -y')
-      box.ssh("rvm install #{@requirements.ruby_version}; echo 'avoiding RVM exit code 5'")
+      box.ssh("rvm install #{@requirements.ruby_version}", :ignore_exit_code => true)
       box.ssh("rvm use #{@requirements.ruby_version} --default")
     end
 
+    private
+
+    def already_installed?(box)
+      box.ssh('rvm current', :capture => true, :ignore_exit_code => true) == @requirements.ruby_version
+    end
   end
 
 end
